third_party/WebKit/LayoutTests/inspector/storage-panel-dom-storage-undo-redo-expected.txt

This test checks the undo/redo operations are performed correctly on the DOM storage views


Running: initialize
Initialized localStorage and sessionStorage by clearing entries.

Running: undoLocalStorageWithEmptyStack
LocalStorage contents:[]

Running: redoLocalStorageWithEmptyStack
LocalStorage contents:[]

Running: localStorageUndoInterlacedWithAddition
LocalStorage contents:[a1=b1]
LocalStorage contents:[a1=b1,a3=b3]
LocalStorage contents:[a1=b1]

Running: addLocalStorageEntries
LocalStorage contents:[a1=b1,a2=b2]

Running: undoLocalStorageLastAddition
LocalStorage contents:[a1=b1]

Running: undoSessionStorageWithEmptyStack
SessionStorage contents:[]

Running: redoSessionStorageWithEmptyStack
SessionStorage contents:[]

Running: undoLocalStorageBeyondBounds
The entry a1=b1 is removed and any attempt to undo beyond it shouldn't cause any failure!
LocalStorage contents:[]

Running: addSessionStorageEntries
SessionStorage contents:[p1=q1,p2=q2,p3=q3,p4=q4]

Running: redoLocalStorageBeyondBounds
The entry a1=b1 and a2=b2 is added back and any attempt to redo beyond it shouldn't cause any failure!
LocalStorage contents:[a1=b1,a2=b2]

Running: undoSessionStorageLastAddition
SessionStorage contents:[p1=q1,p2=q2,p3=q3]

Running: modifyLocalStorageValues
LocalStorage contents:[a1=x1,a2=x2]

Running: undoLocalStorageModifications
LocalStorage contents:[a1=b1,a2=b2]

Running: redoSessionStorageLastAddition
SessionStorage contents:[p1=q1,p2=q2,p3=q3,p4=q4]

Running: redoLocalStorageModifications
LocalStorage contents:[a1=x1,a2=x2]

Running: modifySessionStorageEntriesKey
SessionStorage contents:[m1=q1,m2=q2,m3=q3,m4=q4]

Running: undoLocalStorageModifications
LocalStorage contents:[a1=b1,a2=b2]

Running: undoSessionStorageKeyModifications
SessionStorage contents:[p1=q1,p2=q2,p3=q3,p4=q4]

Running: validateMaxUndoRedoStackDepth
Undo/redo stack depth limit is 20
Performing 40 actions
LocalStorage contents:[a1=b1,a2=b2,x10=y10,x11=y11,x12=y12,x13=y13,x14=y14,x15=y15,x16=y16,x17=y17,x18=y18,x19=y19,x1=y1,x20=y20,x2=y2,x3=y3,x4=y4,x5=y5,x6=y6,x7=y7,x8=y8,x9=y9]
Performing undo operation three times the depth i.e. 60 times but only last 20 operations are undone
LocalStorage contents:[a1=b1,a2=b2,x10=y10,x1=y1,x2=y2,x3=y3,x4=y4,x5=y5,x6=y6,x7=y7,x8=y8,x9=y9]

Running: clearDOMStorageAndFinish
Finished DOMStorage undo/redo tests and cleared localStorage and sessionStorage entries.
